# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMPADU Keuangan API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/keuangan/status-pembayaran": {
            "get": {
                "tags": [
                    "Status Pembayaran"
                ],
                "summary": "Ambil semua status pembayaran",
                "operationId": "63af60d32c44bbd2d699017bc3f6a9c6",
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StatusPembayaran"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Status Pembayaran"
                ],
                "summary": "Buat status pembayaran baru",
                "operationId": "32717b12bf3e646b6593ac7d9a26471c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nim",
                                    "kategori_ukt_id",
                                    "status"
                                ],
                                "properties": {
                                    "nim": {
                                        "type": "string",
                                        "example": "220501001"
                                    },
                                    "kategori_ukt_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Lunas",
                                            "Belum Bayar"
                                        ],
                                        "example": "Lunas"
                                    },
                                    "tanggal_bayar": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusPembayaran"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
        "/api/keuangan/status-pembayaran/{id}": {
            "get": {
                "tags": [
                    "Status Pembayaran"
                ],
                "summary": "Detail status pembayaran",
                "operationId": "ab2a19417f3a3ca207871b5812465e60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusPembayaran"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Status Pembayaran"
                ],
                "summary": "Update status pembayaran",
                "operationId": "1baf66815b2cf836a9174053a8666bec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Lunas",
                                            "Belum Bayar"
                                        ]
                                    },
                                    "tanggal_bayar": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusPembayaran"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Status Pembayaran"
                ],
                "summary": "Hapus status pembayaran",
                "operationId": "13ebb0160555001662cdce58b9530a3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StatusPembayaran": {
                "required": [
                    "kode",
                    "nama"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "kode": {
                        "type": "string",
                        "example": "LUNAS"
                    },
                    "nama": {
                        "type": "string",
                        "example": "Lunas"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-01T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-01T08:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}